requirements.txt
matplotlib
python-dotenv
textblob
azure-functions
azure-ai-projects
azure-identity
